per PR comment: move Python 2 `is_identifier` replacement from `_main_legacy_py.py` to `main.py`
