+++ src/makefun/main.py
@@ -16,6 +16,19 @@ from textwrap import dedent
 from types import FunctionType
 
 
+if sys.version_info >= (3, 0):
+    is_identifier = str.isidentifier
+else:
+    def is_identifier(string):
+        """
+        Replacement for `str.isidentifier` when it is not available (e.g. on Python 2).
+        :param string:
+        :return:
+        """
+        if len(string) == 0 or string[0].isdigit():
+            return False
+        return all([s.isalnum() for s in string.split("_")])
+
 try:  # python 3.3+
     from inspect import signature, Signature, Parameter
 except ImportError:
@@ -1558,9 +1571,3 @@ def compile_fun_manually(target,
     new_f.__source__ = source_lines
 
     return new_f
-
-
-if sys.version_info >= (3, 0):
-    is_identifier = str.isidentifier
-else:
-    from makefun._main_legacy_py import is_identifier
